rust: add option to generate BTreeMap/Set code and tune traits accordingly
HashMap/HashSet do not implement Hash, and so they cannot themselves
be keys/members of a HashMap/Set. BTreeMap/Set implement Ord, and
so they can be keys/members of a BTreeMap/Set.

Since in general HashMap/Set is preferred, its the default. However,
if you have a Thrift spec with complex nesting of maps and sets
then the btree_mapset option will generate code to use BTrees.

This also means that the set of derived traits for generated types
depends on both the recursive membership of structures and other
aggregate types, as well as whether they're Hash or BTree implementations.
The generated code adds the largest set of derives possible for each
type.